demo github

+++ mobil.h
@@ -27,11 +27,12 @@ struct brand {
 
 struct product {
     string name;
-    // Up to 1 quintillion (10^18)
-    unsigned long long int price;
+    unsigned long int price;
     string type;
     adr_product next;
 };
+
+
 
 struct mobil {
     adr_brand first;
